+++ example/a.ts
@@ -1,19 +1,16 @@
 import { Elysia, t } from '../src'
 import { post, req } from '../test/utils'
 
-const model = new Elysia()
-	.model({
-		number: t.Number({ default: 0 })
-	})
+const model = new Elysia().model({
+	number: t.Number({ default: 0 })
+})
 
-const app = new Elysia()
-	.use(model)
-	.post('/', ({ body }) => body, {
-		body: t.Object({
-			name: t.String(),
-			amount: t.Optional(model.Ref('number'))
-		})
+const app = new Elysia().use(model).post('/', ({ body }) => body, {
+	body: t.Object({
+		name: t.String(),
+		age: t.Optional(model.Ref('number'))
 	})
+})
 
 const result = await app
 	.handle(
@@ -24,17 +21,3 @@ const result = await app
 	.then((x) => x.json())
 
 console.log(result)
-
-// console.dir(
-// 	t
-// 		.Module({
-// 			number: t.Numeric(),
-// 			optionalNumber: t.Optional(t.Ref('number'))
-// 		})
-// 		.Import('optionalNumber'),
-// 	{
-// 		depth: null
-// 	}
-// )
-
-// console.log(correct)